moved openshift platform login page into login folder Signed-off-by: mdolhalo <mdolhalo@redhat.com>
