+++ tests/e2e/index.ts
@@ -30,14 +30,13 @@ export * from './pageobjects/login/interfaces/ICheLoginPage';
 export * from './pageobjects/login/interfaces/IOcpLoginPage';
 export * from './pageobjects/login/kubernetes/DexLoginPage';
 export * from './pageobjects/login/kubernetes/KubernetesLoginPage';
+export * from './pageobjects/login/openshift/OcpLoginPage';
 export * from './pageobjects/login/openshift/OcpRedHatLoginPage';
 export * from './pageobjects/login/openshift/OcpUserLoginPage';
 export * from './pageobjects/login/openshift/RedHatLoginPage';
 export * from './pageobjects/login/openshift/RegularUserOcpCheLoginPage';
-export * from './pageobjects/openshift/CheLoginPage';
 export * from './pageobjects/openshift/OcpApplicationPage';
 export * from './pageobjects/openshift/OcpImportFromGitPage';
-export * from './pageobjects/openshift/OcpLoginPage';
 export * from './pageobjects/openshift/OcpMainPage';
 export * from './tests-library/LoginTests';
 export * from './tests-library/ProjectAndFileTests';
